리포트에서 BMI와 건강상태, 기록한 몸무게 목록 표시

+++ src/main/java/com/project/cobell/dto/UserDto.java
@@ -11,8 +11,8 @@ public class UserDto {
 	private int gender;
 	private int age;
 	private int goal;
-	private int height;
-	private int weight;
+	private float height;
+	private float weight;
 	private String title;
 	private String content;
 }
